# web-pilot/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperatorOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disabled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locked_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Until"
          },
          "password_changed_at": {
            "format": "date-time",
            "title": "Password Changed At",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "created_at",
          "disabled_at",
          "password_changed_at",
          "locked_until"
        ],
        "title": "OperatorOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "courier API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operators": {
      "get": {
        "operationId": "list_operators_operators_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OperatorOut"
                  },
                  "title": "Response List Operators Operators Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Operators"
      },
      "post": {
        "operationId": "create_operator_operators_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Operator"
      }
    },
...
    "/pilots": {
      "get": {
        "operationId": "list_pilots_pilots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PilotOut"
                  },
                  "title": "Response List Pilots Pilots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Pilots"
      },
      "post": {
        "operationId": "create_pilot_pilots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pilot"
      }
    },
...
    "/uas/aircraft": {
      "get": {
        "description": "`q` matches serial, label or model. Fleet aircraft with no\noperator are included only with `include_fleet`.",
        "operationId": "list_uas_uas_aircraft_get",
        "parameters": [
          {
            "in": "query",
            "name": "operator_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Operator Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegistrationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "class_label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ClassLabel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Class Label"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "include_fleet",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Fleet",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UasOut"
                  },
                  "title": "Response List Uas Uas Aircraft Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Uas",
        "tags": [
          "uas registry"
        ]
      },
      "post": {
        "operationId": "register_uas_uas_aircraft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UasIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Uas",
        "tags": [
          "uas registry"
        ]
      }
    },
    "/uas/aircraft/lookup": {
      "get": {
        "description": "Exactly one aircraft, by serial: an exact match, else the only one\nthat matches ignoring case.",
        "operationId": "uas_by_serial_uas_aircraft_lookup_get",
        "parameters": [
          {
            "in": "query",
            "name": "serial",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Serial",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Uas By Serial",
        "tags": [
          "uas registry"
        ]
      }
    },
    "/uas/aircraft/{drone_id}": {
      "get": {
        "operationId": "get_uas_uas_aircraft__drone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Uas",
        "tags": [
          "uas registry"
        ]
      },
      "patch": {
        "operationId": "update_uas_uas_aircraft__drone_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UasPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Uas",
        "tags": [
          "uas registry"
        ]
      }
    },
...
    "/uas/operators": {
      "get": {
        "description": "`q` matches the registration number or the legal name.",
        "operationId": "list_operators_uas_operators_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RegistrationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecordSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UasOperatorOut"
                  },
                  "title": "Response List Operators Uas Operators Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Operators",
        "tags": [
          "uas registry"
        ]
      },
      "post": {
        "operationId": "create_operator_uas_operators_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UasOperatorIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOperatorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Operator",
        "tags": [
          "uas registry"
        ]
      }
    },
    "/uas/operators/lookup": {
      "get": {
        "description": "Exactly one operator, by registration number, ignoring case.",
        "operationId": "operator_by_registration_uas_operators_lookup_get",
        "parameters": [
          {
            "in": "query",
            "name": "registration_number",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Registration Number",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UasOperatorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator By Registration",
        "tags": [
          "uas registry"
        ]
      }
    },
...
    "/uas/pilots/{pilot_id}": {
      "get": {
        "operationId": "get_pilot_uas_pilots__pilot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pilot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pilot Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemotePilotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pilot",
        "tags": [
          "uas registry"
        ]
      }
    },
    "/uas/pilots/{pilot_id}/competencies": {
      "put": {
        "description": "Record one competency, replacing the pilot's previous record of it.",
        "operationId": "record_competency_uas_pilots__pilot_id__competencies_put",
        "parameters": [
          {
            "in": "path",
            "name": "pilot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pilot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemotePilotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Competency",
        "tags": [
          "uas registry"
        ]
      }
    },
...
  }
}